+++ src/main/java/lotto/domain/LottoPrizeMap.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.enums.LottoPrize;
