+++ src/main/java/lotto/service/LottoDraw.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.Lotto;
 
 import java.util.ArrayList;
 import java.util.Collections;
