+++ src/main/java/lotto/view/InputView.java
@@ -1,8 +1,8 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.Lotto;
-import lotto.Validate;
+import lotto.domain.Lotto;
+import lotto.util.InputValidator;
 import lotto.enums.InputMessage;
 
 import java.util.Arrays;
@@ -15,9 +15,9 @@ public class InputView {
         System.out.println(InputMessage.PURCHASE_AMOUNT_INPUT.getMessage());
         String purchaseAmount = Console.readLine();
 
-        Validate.validateInteger(purchaseAmount);
-        Validate.validatePurchaseAmountPositive(Integer.parseInt(purchaseAmount));
-        Validate.validatePurchaseAmountUnit(Integer.parseInt(purchaseAmount));
+        InputValidator.validateInteger(purchaseAmount);
+        InputValidator.validatePurchaseAmountPositive(Integer.parseInt(purchaseAmount));
+        InputValidator.validatePurchaseAmountUnit(Integer.parseInt(purchaseAmount));
 
         return Integer.parseInt(purchaseAmount);
     }
@@ -26,8 +26,8 @@ public class InputView {
         System.out.println(InputMessage.PRIZE_NUMBERS_INPUT.getMessage());
         List<String> lottoNumbers = Arrays.stream(Console.readLine().split(LOTTO_NUMBERS_SPLIT_STRING)).toList();
 
-        Validate.validateNumbersInteger(lottoNumbers);
-        Validate.validateLottoNumbersRange(lottoNumbers);
+        InputValidator.validateNumbersInteger(lottoNumbers);
+        InputValidator.validateLottoNumbersRange(lottoNumbers);
 
         return lottoNumbers.stream()
                 .map(Integer::parseInt)
@@ -38,9 +38,9 @@ public class InputView {
         System.out.println(InputMessage.BONUS_NUMBER_INPUT.getMessage());
         String bonusNumber = Console.readLine();
 
-        Validate.validateInteger(bonusNumber);
-        Validate.validateLottoNumberRange(Integer.parseInt(bonusNumber));
-        Validate.validateLottoBonusNumberDuplication(winningNumbers, Integer.parseInt(bonusNumber));
+        InputValidator.validateInteger(bonusNumber);
+        InputValidator.validateLottoNumberRange(Integer.parseInt(bonusNumber));
+        InputValidator.validateLottoBonusNumberDuplication(winningNumbers, Integer.parseInt(bonusNumber));
 
         return Integer.parseInt(bonusNumber);
     }
